Moved general functions to utils submodule
